+++ app/components/Entity/components/ConfigureWorkflowInputs/components/Main/components/Stepper/components/Step/GTFStep/step.ts
@@ -1,0 +1,9 @@
+import { StepConfig } from "../types";
+import { GTFStep } from "./gtfStep";
+
+export const STEP: StepConfig = {
+  Step: GTFStep,
+  description: "Select the GTF files for the workflow",
+  key: "geneModelUrl",
+  label: "GTF Files",
+};