+++ app/components/Entity/components/ConfigureWorkflowInputs/components/Main/components/Stepper/components/Step/ReferenceAssemblyStep/referenceAssemblyStep.tsx
@@ -1,0 +1,31 @@
+import { Step } from "@databiosphere/findable-ui/lib/components/Stepper/components/Step/step";
+import { StepContent } from "@databiosphere/findable-ui/lib/components/Stepper/components/Step/components/StepContent/stepContent";
+import { StepLabel } from "@databiosphere/findable-ui/lib/components/Stepper/components/Step/components/StepLabel/stepLabel";
+import { StepProps } from "../types";
+import { Optional } from "@databiosphere/findable-ui/lib/components/Stepper/components/Step/components/StepLabel/components/Optional/optional";
+import { useEffect } from "react";
+
+export const ReferenceAssemblyStep = ({
+  active,
+  completed,
+  entryKey,
+  entryLabel,
+  genome,
+  index,
+  onConfigure,
+}: StepProps): JSX.Element => {
+  const { accession } = genome;
+
+  useEffect(() => {
+    onConfigure(entryKey, entryLabel, [{ key: accession, value: accession }]);
+  }, [accession, entryKey, entryLabel, onConfigure]);
+
+  return (
+    <Step active={active} completed={completed} index={index}>
+      <StepLabel optional={<Optional>{accession}</Optional>}>
+        {entryLabel}
+      </StepLabel>
+      <StepContent>None</StepContent>
+    </Step>
+  );
+};