+++ app/components/Entity/components/ConfigureWorkflowInputs/components/Main/components/Stepper/components/Step/ReferenceAssemblyStep/step.ts
@@ -1,0 +1,9 @@
+import { StepConfig } from "../types";
+import { ReferenceAssemblyStep } from "./referenceAssemblyStep";
+
+export const STEP: StepConfig = {
+  Step: ReferenceAssemblyStep,
+  disabled: true,
+  key: "referenceAssembly",
+  label: "Reference Assembly",
+};