+++ app/components/Entity/components/ConfigureWorkflowInputs/components/Main/components/Stepper/components/Step/components/Button/constants.ts
@@ -1,0 +1,10 @@
+import {
+  COLOR,
+  VARIANT,
+} from "@databiosphere/findable-ui/lib/styles/common/mui/button";
+import { ButtonProps } from "@mui/material";
+
+export const BUTTON_PROPS: ButtonProps = {
+  color: COLOR.PRIMARY,
+  variant: VARIANT.CONTAINED,
+};