+++ app/components/Entity/components/ConfigureWorkflowInputs/components/Main/components/Stepper/stepper.styles.ts
@@ -1,0 +1,6 @@
+import styled from "@emotion/styled";
+import { Stepper } from "@mui/material";
+
+export const StyledStepper = styled(Stepper)`
+  gap: 16px;
+`;