+++ app/components/Entity/components/ConfigureWorkflowInputs/components/Main/types.ts
@@ -1,0 +1,15 @@
+import {
+  BRCDataCatalogGenome,
+  Workflow,
+} from "../../../../../../apis/catalog/brc-analytics-catalog/common/entities";
+import {
+  ConfiguredInput,
+  OnConfigure,
+} from "../../../../../../views/WorkflowInputsView/hooks/UseConfigureInputs/types";
+
+export interface Props {
+  configuredInput: ConfiguredInput;
+  genome: BRCDataCatalogGenome;
+  onConfigure: OnConfigure;
+  workflow: Workflow;
+}